==> 1-data/1-notebooks/01-misc.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---
```{r misc-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(magrittr)
library(operator.tools)
library(knitr)
library(rprojroot)
library(tidyverse)
library(rgdal)
library(sp)
library(rgeos)
library(miscgis)
library(tigris)
library(leaflet)
library(ggthemes)
library(stringr)
library(downloader)
library(miscgis)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
options(tigris_class = "sf")
```

```{r misc-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
```

Below is a collection of spatial datasets used in this project.

### King County Boundary  {-}
```{r misc-kc, fig.cap="King County\'s geographic boundary"}

if(!file.exists(root_file('./1-data/4-interim/kc-noclip-sf.rds'))){
        tigris::counties(state = "53") %>% 
                st_transform(crs_proj@projargs) %>% 
                filter(NAME %in% 'King') %>% 
                write_rds(root_file('./1-data/4-interim/kc-noclip-sf.rds'))
}

kc_noclip_sf <- read_rds(root_file('./1-data/4-interim/kc-noclip-sf.rds'))

myLfltGrey(data = as(kc_noclip_sf,'Spatial')) %>% 
        myLfltOpts() %>% 
        addPolygons(color = red,opacity = 1,fillColor = red,fillOpacity = .5)
```


### King County Subdivision Boundary {-}

```{r misc-sea-ccd, fig.cap="Seattle Subdivision of King County\'s geographic boundary"}

if(!file.exists(root_file('1-data/4-interim/seattle-ccd-noclip-sf.rds'))){
        tigris::county_subdivisions(state = "53",county = "033") %>% 
                st_transform(crs_proj@projargs) %>% 
                filter(NAME %in% "Seattle") %>% 
                mutate(NAME = 'Seattle CCD') %>% 
                write_rds(root_file('1-data/4-interim/seattle-ccd-noclip-sf.rds'))
}

sea_ccd_noclip_sf <- read_rds(root_file('1-data/4-interim/seattle-ccd-noclip-sf.rds'))

myLfltGrey(data = as(sea_ccd_noclip_sf,'Spatial')) %>% 
        myLfltOpts() %>% 
        addPolygons(color = blue,opacity = 1,fillColor = blue,fillOpacity = .5)

```


### Seattle Boundary {-}

```{r misc-sea-bound, fig.cap="Seattle\'s geographic boundary"}

if(!file.exists(root_file('1-data/4-interim/seattle-noclip-sf.rds'))){
        tigris::places(state = 53) %>%
                tigris::filter_place(place = "Seattle") %>%
                st_transform(crs_proj@projargs) %>% 
                write_rds(root_file('1-data/4-interim/seattle-noclip-sf.rds'))
        
}

sea_noclip_sf <- read_rds(root_file('1-data/4-interim/seattle-noclip-sf.rds'))

myLfltGrey(data = as(sea_noclip_sf,'Spatial')) %>% 
        myLfltOpts() %>% 
        addPolygons(color = green,opacity = 1,fillColor = green,fillOpacity = .5)

```


### Tracts in King County {-}
Although this assessment is primarily focused on three communities within the Seattle CCD subdivision of King County, one of the indicators (housing market conditions) uses neighboring tracts to determine displacement risk. Some of the neighboring tracts are part of other county subdivision, but rather than targeting just those specific tracts, this method collects data for all King County tracts and then runs the analysis on the appropriate subsets.

In the absence of a straight-forward method for identifying all the census tracts in the Seattle CCD subdivision of King County, it is possible to extract this information from [American Factfinder](https://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml). This [tutorial](https://ask.census.gov/prweb/PRServletCustom?pyActivity=pyMobileSnapStart&Action=showHarness&Purpose=KMHelpSitePortal&className=Data-Portal&ReadOnly=true&pzMobileInitActivity=KMFetchArticleExternal&pzMobileInitActivityParams=%26ArticleID%3DKCP-3256&pzMobileContextPageName=pyDisplayHarness) describes how to use the American Factfinder interface to extract a list of all "all tracts within (or partially within) a census place"; substituting "county subdivision" for "place" will retrieve the desired results.

Tract boundaries sometimes change over time, particularly on years when the decennial census is conducted. While this project will primarily use the current set of boundaries, it is necessary to collect the tract boundaries of the 2000-2009 time period so that these datasets can be visually represented in choropleth diagrams.

```{r misc-tr-kc-noclip, fig.height=2, fig.width=2, fig.show='hold', dpi=150}
# 2015 tract boundaries
if(!file.exists(root_file('1-data/4-interim/tr-kc-noclip-2015-sf.rds'))){
        
        # Seattle CCD tracts 
        # Note: because this selection includes all tracts "within or partially-within" the Seattle CC,
        # several tract GEOIDs are duplicated in the selection. For the sake of clarity, these duplicates
        # are removed from the final vector of GEOIDs.
        tr_ccd_geoid_2012 <- 
                read_csv(
                        root_file('1-data/3-external/manual/seattle-ccd/ACS_12_5YR_B01001/ACS_12_5YR_B01001_with_ann.csv'), 
                        col_types = cols(Id2 = col_character()), 
                        skip = 1) %>% 
                mutate(NEW_GEOID1 = str_sub(Id2,1,5),
                       NEW_GEOID2 = str_sub(Id2,16,21),
                       GEOID = paste0(NEW_GEOID1,NEW_GEOID2),
                       UNIQUE = !duplicated(GEOID)) %>%
                filter(UNIQUE) %>% 
                extract2('GEOID')
        
        
        # All 2015 KC tracts with Seattle CCD subdivision column
        
        tigris::tracts(state = '53',county = '033', year = 2015) %>%
                st_transform(crs_proj@projargs) %>% 
                mutate(SEACCD_LGL = ifelse(GEOID %in% tr_ccd_geoid_2012,TRUE,FALSE)) %>% 
                write_rds(root_file('1-data/4-interim/tr-kc-noclip-2015-sf.rds'))
              
}

# pre-2010 tract boundaries
if(!file.exists(root_file('1-data/4-interim/tr-kc-noclip-2009-sf.rds'))){
        
        # Seattle CCD tracts 
        # Note: because this selection includes all tracts "within or partially-within" the Seattle CC,
        # several tract GEOIDs are duplicated in the selection. For the sake of clarity, these duplicates
        # are removed from the final vector of GEOIDs.
        tr_ccd_geoid_2009 <- 
                read_csv(
                        root_file('1-data/3-external/manual/seattle-ccd/ACS_09_5YR_B01001/ACS_09_5YR_B01001_with_ann.csv'), 
                        col_types = cols(GEO.id2 = col_character())) %>% 
                mutate(NEW_GEOID1 = str_sub(GEO.id2,1,5),
                       NEW_GEOID2 = str_sub(GEO.id2,16,21),
                       GEOID = paste0(NEW_GEOID1,NEW_GEOID2),
                       UNIQUE = !duplicated(GEOID)) %>%
                filter(UNIQUE) %>% 
                extract2('GEOID')
        
        
        # All 2015 KC tracts with Seattle CCD subdivision column
        
        my_dl_zip <- function(url,dir_filepath){
                temp <- tempfile()
                downloader::download(url, dest = temp, mode='wb')
                unzip(temp, exdir = dir_filepath)
                
        }
        
        # my_dl_zip('ftp://ftp.census.gov/geo/tiger/TIGER2009/53_WASHINGTON/53033_King_County/tl_2009_53033_tract00.zip',root_file('1-data/3-external/manual/kc-tr/'))


        # If the download script (above) doesn't work, open this url (https://www.census.gov/cgi-bin/geo/shapefiles2009/county-files?county=53033)
        # and download the 'Census Tract (Census 2000)' file 
        
        readOGR(dsn = root_file('./1-data/3-external/manual/kc-tr/tl_2009_53033_tract00/'),
                layer = 'tl_2009_53033_tract00',
                verbose = FALSE,
                stringsAsFactors = FALSE) %>% 
                st_as_sf() %>% 
                st_transform(crs_proj@projargs) %>% 
                st_cast('MULTIPOLYGON') %>%  
                mutate(SEACCD_LGL = ifelse(CTIDFP00 %in% tr_ccd_geoid_2009,TRUE,FALSE)) %>% 
                write_rds(root_file('1-data/4-interim/tr-kc-noclip-2009-sf.rds'))
                
        
        # Note: the `tigris` package is returning an error for attempts to
        # download tract data before 2011
        # tigris::tracts(state = '53',county = '033', year = 2009) %>%
        #         spTransform(CRSobj = crs_proj) %>%
        #         st_as_sf() %>%
        #         st_cast('MULTIPOLYGON') %>%
        #         mutate(SEACCD_LGL = ifelse(GEOID %in% tr_ccd_geoid_2012,TRUE,FALSE)) %>%
        #         write_rds(root_file('1-data/4-interim/tr-kc-wtr-2009-sf.rds'))
        
}


tr_kc_noclip_2015_sf <- read_rds(root_file('1-data/4-interim/tr-kc-noclip-2015-sf.rds'))

tr_kc_noclip_2009_sf <- read_rds(root_file('1-data/4-interim/tr-kc-noclip-2009-sf.rds'))

show_tr_ccd_noclip_2009_sf <- function(){
        
        blue_orange <- c(blue,orange) %>% unlist
        
        pal <- colorFactor(blue_orange,levels = c(TRUE,FALSE), ordered = TRUE)
        
        myLfltGrey() %>% 
                myLfltOpts() %>% 
                addPolygons(data = as(tr_kc_noclip_2009_sf,"Spatial"),
                            weight = .5,
                            color = ~ pal(SEACCD_LGL),
                            opacity = 1,
                            fillColor = ~ pal(SEACCD_LGL),
                            fillOpacity = .5) %>% 
                addLegend(position = 'bottomright',colors = blue_orange,labels = c('Seattle CCD','Other Tracts'),title = 'pre-2010') %>% 
                miscgis::styleWidget(style = "float:left;margin:1px")
}
show_tr_ccd_noclip_2015_sf <- function(){
        
        blue_orange <- c(blue,orange) %>% unlist
        
        pal <- colorFactor(blue_orange,levels = c(TRUE,FALSE), ordered = TRUE)
        
        myLfltGrey() %>% 
                myLfltOpts() %>% 
                addPolygons(data = as(tr_kc_noclip_2015_sf,"Spatial"),
                            weight = .5,
                            color = ~ pal(SEACCD_LGL),
                            opacity = 1,
                            fillColor = ~ pal(SEACCD_LGL),
                            fillOpacity = .5) %>% 
                addLegend(position = 'bottomright',colors = blue_orange,labels = c('Seattle CCD','Other Tracts'),title = '2010') %>% 
                miscgis::styleWidget(style = "float:none;margin:1px")
}


show_tr_ccd_noclip_2009_sf()
show_tr_ccd_noclip_2015_sf()


```

### King County Waterbodies {-}

These are useful for "clipping" census geographies whose boundaries extend into waterbodies.

```{r misc-wtr, fig.cap="Puget Sound waterbodies"}

if(!file.exists(root_file('1-data/4-interim/wtr-sf.rds'))){
        fp_wtr <- root_file('1-data/3-external/NHDMajor.gdb')

# check if the file already exists, if not then download it
if(!file.exists(fp_wtr)){
        
        url <- "ftp://www.ecy.wa.gov/gis_a/inlandWaters/NHD/NHDmajor.gdb.zip" # save the URL for the waterbodies data
        
        temp <- tempfile() # create a temporary file to hold the compressed download
        
        download(url, dest = temp, mode="wb") # download the file
        
        unzip (temp, exdir = root_file('1-data/3-external/')) # extract the ESRI geodatabase file to a project folder
}

wtr_sf <-
        suppressWarnings(readOGR(dsn = fp_wtr,      # create a waterbodies shape
                layer = "NHD_MajorWaterbodies",verbose = FALSE,pointDropZ = TRUE)) %>%
        gBuffer(byid=TRUE, width=0) %>% # clean up self-intersecting polygons
        spTransform(CRSobj = crs_proj) %>%   # transform the projection to match the project projection
        st_as_sf() %>% 
        st_cast('MULTIPOLYGON') 

wtr_kc_sf <- 
        wtr_sf %>% 
        st_intersects(x = kc_noclip_sf,y = .) %>% 
        unlist(use.names = F) %>% 
        wtr_sf[.,]

# All King County waterbodies
wtr_kc_sf %>% write_rds(root_file('1-data/4-interim/wtr-kc-sf.rds'))

# All waterbodies near Seattle CCD (the Puget Sound area)
wtr_kc_sf %>% 
        st_intersects(x = sea_ccd_noclip_sf,y = .) %>% 
        unlist(use.names = F) %>% 
        wtr_kc_sf[.,] %>% 
        write_rds(root_file('1-data/4-interim/wtr-puget-sf.rds'))
}

wtr_kc_sf <- read_rds(root_file('1-data/4-interim/wtr-kc-sf.rds'))

wtr_puget_sf <- read_rds(root_file('1-data/4-interim/wtr-puget-sf.rds'))

show_wtr_kc <- function(){
        myLfltGrey(data = as(wtr_kc_sf,'Spatial')) %>% 
                myLfltOpts() %>% 
                addPolygons(color = blue, opacity = 1, 
                            weight = .5, fillColor = blue,fillOpacity = .5)        
}

show_wtr_kc()

```

### Census Geometries Without (Western) Waterbodies {-}

The same four types of cenus geometries shown above (King County, Seattle CCD, City of Seattle, and all King county tracts) with the major waterbodies removed.

```{r misc-clip-geoms}

# Clip King County
if(!file.exists(root_file('1-data/4-interim/kc-sf.rds'))){
        kc_noclip_sf %>% 
                mutate(geometry = st_difference(geometry, st_union(wtr_kc_sf))) %>% 
                write_rds(root_file('1-data/4-interim/kc-sf.rds'))
}

# Clip Seattle CCD
if(!file.exists(root_file('1-data/4-interim/seattle-ccd-sf.rds'))){
        sea_ccd_noclip_sf %>% 
                mutate(geometry = st_difference(geometry, st_union(wtr_kc_sf))) %>% 
                st_cast('MULTIPOLYGON') %>% 
                write_rds(root_file('1-data/4-interim/seattle-ccd-sf.rds'))
}

# Clip Seattle
if(!file.exists(root_file('1-data/4-interim/seattle-sf.rds'))){
        sea_noclip_sf %>% 
                mutate(geometry = st_difference(geometry, st_union(wtr_kc_sf))) %>% 
                st_cast('MULTIPOLYGON') %>% 
                write_rds(root_file('1-data/4-interim/seattle-sf.rds'))
}

# Clip King County tracts (2015)
if(!file.exists(root_file('1-data/4-interim/tr-kc-2015-sf.rds'))){
        tr_kc_noclip_2015_sf %>% 
                filter(TRACTCE %!in% '990100') %>% # remove the Puget Sound tract
                mutate(geometry = st_difference(geometry, st_union(wtr_kc_sf))) %>% 
                st_cast('MULTIPOLYGON') %>% 
                write_rds(root_file('1-data/4-interim/tr-kc-2015-sf.rds'))
}

# Clip King County tracts (2009)
if(!file.exists(root_file('1-data/4-interim/tr-kc-2009-sf.rds'))){
        tr_kc_noclip_2009_sf %>% 
                filter(TRACTCE00 %!in% '990100') %>% # remove the Puget Sound tract
                mutate(geometry = st_difference(geometry, st_union(wtr_kc_sf))) %>% 
                st_cast('MULTIPOLYGON') %>% 
                write_rds(root_file('1-data/4-interim/tr-kc-2009-sf.rds'))
}

```


```{r misc-show-clipped-geoms-1}

kc_sf <- read_rds(root_file('1-data/4-interim/kc-sf.rds'))

sea_ccd_sf <- read_rds(root_file('1-data/4-interim/seattle-ccd-sf.rds'))

sea_sf <- read_rds(root_file('1-data/4-interim/seattle-sf.rds'))

show_kc_sf <- function(){
        myLfltGrey(data = as(kc_sf,'Spatial')) %>% 
                myLfltOpts() %>% 
                addPolygons(color = red,opacity = 1,fillColor = red,fillOpacity = .5)
}

show_sea_ccd_sf <- function(){
        myLfltGrey(data = as(sea_ccd_sf,'Spatial')) %>% 
                myLfltOpts() %>% 
                addPolygons(color = blue,opacity = 1,fillColor = blue,fillOpacity = .5)
}

show_sea_sf <- function(){
        myLfltGrey(data = as(sea_sf,'Spatial')) %>% 
                myLfltOpts() %>% 
                addPolygons(color = green,opacity = 1,fillColor = green,fillOpacity = .5)
}
show_kc_sf()
show_sea_ccd_sf()
show_sea_sf()
```


```{r show-clipped-geoms-2, fig.height=2, fig.width=2, fig.show='hold', dpi=150}

tr_kc_2015_sf <- read_rds(root_file('1-data/4-interim/tr-kc-2015-sf.rds'))

tr_kc_2009_sf <- read_rds(root_file('1-data/4-interim/tr-kc-2009-sf.rds'))

show_tr_kc_2009_sf <- function(){
        
        blue_orange <- c(blue,orange) %>% unlist
        
        pal <- colorFactor(blue_orange,levels = c(TRUE,FALSE), ordered = TRUE)
        
        myLfltGrey() %>% 
                myLfltOpts() %>% 
                addPolygons(data = as(tr_kc_2009_sf,"Spatial"),
                            weight = .5,
                            color = ~ pal(SEACCD_LGL),
                            opacity = 1,
                            fillColor = ~ pal(SEACCD_LGL),
                            fillOpacity = .5) %>% 
                addLegend(position = 'bottomright',colors = blue_orange,labels = c('Seattle CCD','Other Tracts'),title = 'pre-2010') %>% 
                miscgis::styleWidget(style = "float:left;margin:1px")
}

show_tr_kc_2015_sf <- function(){
        
        blue_orange <- c(blue,orange) %>% unlist
        
        pal <- colorFactor(blue_orange,levels = c(TRUE,FALSE), ordered = TRUE)
        
        myLfltGrey() %>% 
                myLfltOpts() %>% 
                addPolygons(data = as(tr_kc_2015_sf,"Spatial"),
                            weight = .5,
                            color = ~ pal(SEACCD_LGL),
                            opacity = 1,
                            fillColor = ~ pal(SEACCD_LGL),
                            fillOpacity = .5) %>% 
                addLegend(position = 'bottomright',colors = blue_orange,labels = c('Seattle CCD','Other Tracts'),title = '2010') %>% 
                miscgis::styleWidget(style = "float:none;margin:1px")
}

show_tr_kc_2009_sf()

show_tr_kc_2015_sf()

```


### Mixed Census Geometries Simple Feature Object {-}

Add a brief explanation of the `sf` package

```{r misc-combined}

if(!file.exists(root_file('./1-data/4-interim/coo-acs-2015-geoms-sf.rds'))){
        list(
                kc_sf %>% 
                        select(geometry) %>% mutate(NAME = 'KC', NAME_FULL = 'King County, Washington',GEOID6 = NA_character_, SEACCD_LGL = NA),
                sea_ccd_sf %>% 
                        select(geometry) %>% mutate(NAME = 'SEACCD', NAME_FULL = 'Seattle CCD, King County, Washington', GEOID6 = NA_character_, SEACCD_LGL = NA),
                tr_kc_2015_sf %>% 
                        select(NAME,NAME_FULL = NAMELSAD,GEOID6 = GEOID,SEACCD_LGL,geometry) %>% mutate(NAME_FULL = paste0(NAME_FULL,', King County, Washington'),GEOID6 = str_sub(GEOID6,6,11))
        ) %>% 
                reduce(rbind.sf) %>%
                write_rds(root_file('./1-data/4-interim/coo-acs-2015-geoms-sf.rds'))
        
}

if(!file.exists(root_file('./1-data/4-interim/coo-acs-2009-geoms-sf.rds'))){
        list(
                kc_sf %>% 
                        select(geometry) %>% mutate(NAME = 'KC', NAME_FULL = 'King County, Washington',GEOID6 = NA_character_, SEACCD_LGL = NA),
                sea_ccd_sf %>% 
                        select(geometry) %>% mutate(NAME = 'SEACCD', NAME_FULL = 'Seattle CCD, King County, Washington', GEOID6 = NA_character_, SEACCD_LGL = NA),
                tr_kc_2009_sf %>% 
                        select(NAME = NAME00, NAME_FULL = NAMELSAD00,GEOID6 = TRACTCE00,SEACCD_LGL,geometry) %>% mutate(NAME_FULL = paste0(NAME_FULL,', King County, Washington'))
        ) %>% 
                reduce(rbind.sf) %>% 
                write_rds(root_file('./1-data/4-interim/coo-acs-2009-geoms-sf.rds'))
        
}


coo_geoms_2015_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-2015-geoms-sf.rds'))

coo_geoms_2009_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-2009-geoms-sf.rds'))

miscgis::print_paged_sf(coo_geoms_2015_sf,n =10)


```
