==> 1-data/1-notebooks/02-census-acs-communities.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r census-acs-comms-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(forcats)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)

```

```{r census-acs-comms-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")

```


### COO Communities {-}

The primary geographic unit of this assessment is the [census tract](https://www.census.gov/geo/reference/gtc/gtc_ct.html). As is the case with many communities, the census geographies do not coincide exactly with the formal geographic boundary of the study's three sites, and should be considered as _spatial approximations_ of these communities.

Listed below are the geographic identifiers of the census tracts that approximate each site.

```{r census-acs-comms-files, results="asis"}

# Census tract boundaries for the 2015 data

# Rainier Valley
rv <- c("53033010001", "53033010300", "53033010401", "53033011001", "53033011002", "53033011101","53033011102", "53033011700", "53033011800", "53033011900")

# White Center
wc <- c("53033026600","53033026700","53033026500","53033026801","53033026802","53033027000")

# SeaTac/Tukwila

stc_hus <- read_csv(root_file('1-data/3-external/manual/seatac-tr/DEC_10_SF1_H1/DEC_10_SF1_H1_with_ann.csv'), 
                    col_types = cols(Id2 = col_character()), skip = 1)

tuk_hus <- read_csv(root_file('1-data/3-external/manual/tukwila-tr/DEC_10_SF1_H1/DEC_10_SF1_H1_with_ann.csv'), 
                    col_types = cols(Id2 = col_character()), 
                    skip = 1)

stc <-  paste(substr(x = stc_hus$Id2,1,2),substr(x = stc_hus$Id2,8,16),sep = "")

tuk <-  paste(substr(x = tuk_hus$Id2,1,2),substr(x = tuk_hus$Id2,8,16),sep = "")

stctuk <- c(stc,tuk)

# Census tract boundaries for the 2009 data

if(!exists('cw')){cw <- read_rds(root_file('1-data/4-interim/cw-2000-2010.rds'))}

get_GEOID2009 <- function(geoids_2015, cw){
        # browser()
        cw %>% 
                filter(GEOID_2010 %in% geoids_2015) %>% 
                filter(WEIGHT > .01) %>%
                extract2('GEOID_2000') %>% unique()
}

if(!file.exists(root_file('1-data/4-interim/coo-comm-tr-df.rds'))){
        list(
                'RV_2009' = get_GEOID2009(rv,cw),
                'RV_2015' = rv,
                'WC_2009' = get_GEOID2009(wc,cw),
                'WC_2015' = wc,
                'STKW_2009' = get_GEOID2009(stctuk,cw),
                'STKW_2015' = stctuk
        ) %>% 
                miscgis::cbind_fill(.) %>%
                as_tibble() %>%
                set_colnames(
                        c(
                                "Rainier Valley<br>(2009)",
                                "Rainier Valley<br>(2015)",
                                "White Center<br>(2009)",
                                "White Center<br>(2015)",
                                "SeaTac/Tukwila<br>(2009)",
                                "SeaTac/Tukwila<br>(2015)"
                        )
                ) %>% 
                write_rds(root_file('1-data/4-interim/coo-comm-tr-df.rds'))
}

read_rds(root_file('1-data/4-interim/coo-comm-tr-df.rds')) %>% 
        knitr::kable(caption = "Census Tract GEOIDs")

```

To create the community approximations, the tract boundaries of each community are merged and each of the census table estimates are aggregated. In addition to combining the estimates, this method also recalculates the standard error for each census table. It should be noted that this method is only valid for census tables representing count data.^[This limitation is made explicit by the `acs` package creator, Ezra Haber Glenn, [here](http://eglenn.scripts.mit.edu/citystate/).]

```{r census-acs-comms-combined-geoms}

if(!exists('acs_norml_sf')){acs_norml_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-norml-sf.rds'))}

# New columns: `COMMUNITY` and `COMMMUNITY_FULL`

rv6 <- map_chr(rv,str_sub,6,11)

wc6 <- map_chr(wc,str_sub,6,11)

stctuk6 <- map_chr(stctuk,str_sub,6,11)

coo6 <- c(rv6,wc6,stctuk6)

# Aggregation function
aggregate_comms <- function(col){
        
        class <- col %>% class() %>% unique %>% extract2(1)
        
        # CHARACTER COLUMNS
        if(class %in% 'character'){NA_character_
        }else 
                
                # LOGICAL COLUMNS 
                if(class %in% 'logical'){ NA
                }else
                        
                        # SFC COLUMNS 
                        if(class %in% 'sfc_MULTIPOLYGON'){ 
                                col %>% 
                                        st_union %>%
                                        st_cast('MULTIPOLYGON') %>% 
                                        st_buffer(.000001) %>% 
                                        st_cast('MULTIPOLYGON')
                                        
                        }else
                                
                                # ACS COLUMNS
                                if(class %in% 'list' & extract2(col,1) %>% class %in% 'acs'){ 
                                        col %>% reduce(suppressWarnings(rbind.acs)) %>% apply(1,sum) %>% list
                                }
}

if(!file.exists(root_file('./1-data/4-interim/acs_norml_all_geoms_sf.rds'))){
        
        acs_norml_comms_sf <- 
                acs_norml_sf %>% 
                mutate(COMMUNITY_FULL = case_when(.$GEOID6 %in% rv6 ~ 'Rainier Valley',
                                                  .$GEOID6 %in% wc6 ~ 'White Center',
                                                  .$GEOID6 %in% stctuk6 ~ 'SeaTac/Tukwila',
                                                  TRUE ~ NA_character_),
                       COMMUNITY = case_when(.$GEOID6 %in% rv6 ~ 'RV',
                                             .$GEOID6 %in% wc6 ~ 'WC',
                                             .$GEOID6 %in% stctuk6 ~ 'STKW',
                                             TRUE ~ NA_character_)) %>% 
                select(NAME,NAME_FULL,GEOID6,COMMUNITY_FULL,COMMUNITY,everything()) %>% 
                ungroup %>% 
                st_as_sf()
        
        comms <-
                acs_norml_comms_sf %>% 
                filter(!is.na(COMMUNITY)) %>% 
                unclass %>% 
                as_tibble %>% 
                slice_rows('COMMUNITY') %>% 
                dmap(aggregate_comms) %>% 
                mutate(geometry = st_sfc(geometry)) %>% 
                st_as_sf()
        
        all_coo <- 
                acs_norml_comms_sf %>% 
                filter(!is.na(COMMUNITY)) %>% 
                unclass %>% 
                as_tibble %>% 
                dmap(aggregate_comms) %>% 
                mutate(geometry = st_sfc(geometry)) %>% 
                st_as_sf()
        
        acs_norml_all_geoms_sf <- 
                list(acs_norml_comms_sf, comms, all_coo) %>% 
                reduce(rbind.sf) %>% 
                mutate(GEOGRAPHY = case_when(is.na(.$NAME) & !is.na(.$COMMUNITY) ~ 'community',
                                             is.na(.$NAME) & is.na(.$COMMUNITY) ~ 'communities (multiple)',
                                             .$NAME %in% 'SEACCD' ~ 'county subdivision',
                                             .$NAME %in% 'KC' ~ 'county',
                                             TRUE ~ 'tract')) %>% 
                mutate(NAME = case_when(
                        is.na(.$NAME) & .$GEOGRAPHY %in% 'communities (multiple)' ~ 'COO',
                        is.na(.$NAME) & .$COMMUNITY %in% 'RV' ~ .$COMMUNITY,
                        is.na(.$NAME) & .$COMMUNITY %in% 'WC' ~ .$COMMUNITY,
                        is.na(.$NAME) & .$COMMUNITY %in% 'STKW' ~ .$COMMUNITY,
                        TRUE ~ .$NAME)) %>% 
                mutate(NAME_FULL = case_when(
                        .$NAME %in% 'COO' ~ 'Communities of Opportunity',
                        .$NAME %in% 'RV' ~ 'Rainier Valley',
                        .$NAME %in% 'WC' ~ 'White Center',
                        .$NAME %in% 'STKW' ~ 'SeaTac/Tukwila',
                        TRUE ~ .$NAME_FULL)) %>% 
                mutate(GEOG_FCT = factor(GEOGRAPHY,
                                         levels = c('tract',
                                                    'community',
                                                    'communities (multiple)',
                                                    'county subdivision',
                                                    'county'),
                                         ordered = TRUE)) %>% 
                mutate(SEACCD_LGL = if_else(NAME %in% c('SEACCD','COO','RV','WC','STKW'),
                                            TRUE,
                                            SEACCD_LGL)) %>% 
                select(NAME:SEACCD_LGL,GEOGRAPHY,GEOG_FCT,geometry,matches('2015'),matches('2009')) 
        
        acs_norml_all_geoms_sf %>% 
                write_rds(root_file('./1-data/4-interim/acs_norml_all_geoms_sf.rds'))
        
}
        
acs_norml_all_geoms_sf <- read_rds(root_file('./1-data/4-interim/acs_norml_all_geoms_sf.rds'))


acs_norml_all_geoms_sf %>% 
        arrange(fct_rev(GEOG_FCT)) %>% 
        select(NAME:GEOG_FCT,geometry,matches('2015'),matches('2009')) %>% 
        st_as_sf() %>% 
        print_paged_sf


```

```{r census-not-acs-all-geoms}
# This saves a version of the acs data that does not use `acs` format
# This will be useful for summary tables and visualizations.

if(!file.exists(root_file('./1-data/5-tidy/coo-census-not-acs-all-sf.rds'))){
        acs_norml_all_geoms_sf %>% 
                unclass %>% 
                as_tibble() %>% 
                gather(TBL,ACS,matches('2015|2009')) %>% 
                mutate(EST = map_dbl(ACS,estimate) %>% round_any(1) %>% as.integer()) %>% 
                mutate(TYPE = 'EST') %>% 
                unite(TBL_EST,TBL,TYPE,sep = '_') %>% 
                spread(TBL_EST,EST) %>% 
                select(-ACS) %>% 
                st_as_sf() %>% 
                write_rds(root_file('./1-data/5-tidy/coo-census-not-acs-all-sf.rds'))
}
```



#### Community Maps {-}
```{r census-acs-comms-map, fig.height=1.25, fig.width=1,  fig.show='hold', dpi=150}

# Print Community Maps

show_comm_map <- function(nm, color, float_dir){
        
        # Map Center
        
        st_centroid_qt <- quietly(st_centroid)
        
        cntr <- acs_norml_all_geoms_sf %>% 
                filter(GEOGRAPHY %in% 'communities (multiple)') %>% 
                st_bbox()
        halfway <- function(max,min){(max-min)/2 + min}
        
        lng <- cntr[['xmin']]
        lat <- halfway(cntr[['ymax']],cntr[['ymin']])
        
        comm_sp <- 
                acs_norml_all_geoms_sf %>% 
                select(everything(),-matches('2015|2009')) %>% 
                filter(GEOGRAPHY %in% 'tract' & COMMUNITY %in% nm) %>% 
                as("Spatial")
        
        comm_cnt_sp <- 
                acs_norml_all_geoms_sf %>% 
                select(everything(),-matches('2015|2009')) %>% 
                filter(GEOGRAPHY %in% 'community' & COMMUNITY %in% nm)  %>% 
                mutate(geometry = st_centroid(geometry)) %>%
                as("Spatial")
        
        comm_bounds_sp <- 
                acs_norml_all_geoms_sf %>% 
                select(everything(),-matches('2015|2009')) %>% 
                filter(GEOGRAPHY %in% 'community' & COMMUNITY %in% nm)  %>%
                as("Spatial") %>% 
                as("SpatialLines")
        
        map <- 
                comm_sp %>% 
                miscgis::myLfltGrey() %>%
                myLfltOpts(tileLabels = FALSE, fullScreenBtn = FALSE, bumpTileLabels = FALSE) %>% 
                addPolygons(smoothFactor = 0,
                            color = col2hex("white"),
                            weight = .5,
                            opacity = 1,
                            fillColor = color,
                            fillOpacity = .65
                ) %>% 
                addPolylines(data = comm_bounds_sp,
                             fillOpacity = 0,
                             color = proj_dark_grey,
                             opacity = 1,
                             weight = 2,
                             smoothFactor = 0) %>%
                addAwesomeMarkers(data = comm_cnt_sp,
                                  options = markerOptions(opacity = 0),
                                  label = ~ NAME_FULL, 
                                  labelOptions = labelOptions(noHide = TRUE,
                                                              offset = c(0,0),
                                                              direction = 'auto', 
                                                              textOnly = TRUE)) %>% 
                setView(lng,lat,zoom = 9) %>% 
                miscgis::styleWidget(style = paste0("float:",float_dir,";margin:1px;","text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF")) 

        return(map)
        }

names <- c('RV','WC','STKW')

colors <- c(blue, green, red)

float_dirs <- c('left','left','none')

map_list <- pmap(list(names, colors, float_dirs),show_comm_map) 

map_list %>% extract2(1)
map_list %>% extract2(2)
map_list %>% extract2(3)



```
