==> 1-data/1-notebooks/02-king-county-parcels.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r kc-prcl-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(maptools)
library(cleangeo)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```


```{r kc-prcl-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
```


#### Calendar Year 2015 / Tax Year 2016 {-}
```{r kc-prcl-2015}

# Read in original parcel data, remove the problem geometries, convert to points
# -----------------------------------------------------------------------------

if(!file.exists(root_file('1-data/4-interim/over-size-limit/prcl_all_2015_pt_sf.rds'))){
        
           
        
        # Convert the original file into sf
        
        if(!file.exists(root_file('./1-data/4-interim/over-size-limit/prcl-all-2015-sf.rds'))){
                # Figure out which featureclass to load
                
                fp_2015 <- root_file('./1-data/3-external/manual/over-size-limit/Year2016.gdb/')
                
                layers <- ogrListLayers(fp_2015)
                
                layers[1:length(layers)]
                
                # It looks like 'parcel_area' is the spatial layer 
                
                readOGR(dsn = fp_2015,layer = 'parcel_area',verbose = FALSE,stringsAsFactors = FALSE) %>%  
                        st_as_sf() %>% 
                        st_transform(crs_proj@projargs) %>% 
                        write_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-2015-sf.rds'))
                
        }
        
        # Read in the sf object
        prcl_all_2015_sf <- read_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-2015-sf.rds'))
        
        
        # Identify the trouble polygons
        if(!file.exists(root_file('./1-data/4-interim/prcl-2015-bad-geoms.rds'))){
                prcl_all_2015_sf %>% 
                        as('Spatial') %>% 
                        clgeo_CollectionReport() %>% 
                        clgeo_SuspiciousFeatures() %>% 
                        write_rds(root_file('./1-data/4-interim/prcl-2015-bad-geoms.rds'))
        }
        
        bad_geom_indices <- read_rds(root_file('./1-data/4-interim/prcl-2015-bad-geoms.rds'))
                
        
        # Remove the trouble polygons,
        # tranform the geometries from polygon to point with `st_centroid()`,
        # and reproject
        prcl_all_2015_pt_sf <- 
                prcl_all_2015_sf[1:nrow(prcl_all_2015_sf) %!in% bad_geom_indices,] %>% 
                st_centroid() %>% 
                st_transform(crs_proj@projargs)
        
        # Save the sf object
        prcl_all_2015_pt_sf %>% write_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_2015_pt_sf.rds'))
}

prcl_all_2015_pt_sf <- read_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_2015_pt_sf.rds'))

# Print a random selection from the data set

prcl_all_2015_pt_sf %>% miscgis::print_paged_sf()


```


#### Calendar Year 2004 / Tax Year 2005 {-}
```{r kc-prcl-2004}

# Read in original parcel data, remove the problem geometries, convert to points
# -----------------------------------------------------------------------------

if(!file.exists(root_file('1-data/4-interim/over-size-limit/prcl_all_2004_pt_sf.rds'))){
        
           
        
        # Convert the original file into sf
        
        if(!file.exists(root_file('./1-data/4-interim/over-size-limit/prcl-all-2004-sf.rds'))){
                # Figure out which featureclass to load
                
                fp_2004 <- root_file('./1-data/3-external/manual/over-size-limit/Year2005.gdb/')
                
                # layers <- ogrListLayers(fp_2004)
                # 
                # layers[1:length(layers)]
                
                # It looks like 'parcel_area' is the spatial layer 
                
                readOGR(dsn = fp_2004,layer = 'parcel_area',verbose = FALSE,stringsAsFactors = FALSE) %>% 
                        st_as_sf() %>% 
                        st_transform(crs_proj@projargs) %>% 
                        write_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-2004-sf.rds'))
                
        }
        
        # Read in the sf object
        prcl_all_2004_sf <- read_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-2004-sf.rds'))
        
        
        # Identify the trouble polygons
        if(!file.exists(root_file('./1-data/4-interim/prcl-2004-bad-geoms.rds'))){
                prcl_all_2004_sf %>% 
                        as('Spatial') %>% 
                        clgeo_CollectionReport() %>% 
                        clgeo_SuspiciousFeatures() %>% 
                        write_rds(root_file('./1-data/4-interim/prcl-2004-bad-geoms.rds'))
        }
        
        bad_geom_indices <- read_rds(root_file('./1-data/4-interim/prcl-2004-bad-geoms.rds'))
                
        
        # Remove the trouble polygons,
        # tranform the geometries from polygon to point with `st_centroid()`,
        # and reproject
        prcl_all_2004_pt_sf <- 
                prcl_all_2004_sf[1:nrow(prcl_all_2004_sf) %!in% bad_geom_indices,] %>% 
                st_centroid() %>% 
                st_transform(crs_proj@projargs)
        
        # Save the sf object
        prcl_all_2004_pt_sf %>% write_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_2004_pt_sf.rds'))
}

prcl_all_2004_pt_sf <- read_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_2004_pt_sf.rds'))

prcl_all_2004_pt_sf %>% miscgis::print_paged_sf()

```


#### Calendar Year 1999 / Tax Year 2000 {-}
```{r kc-prcl-1999}

# Read in original parcel data, remove the problem geometries, convert to points
# -----------------------------------------------------------------------------

if(!file.exists(root_file('1-data/4-interim/over-size-limit/prcl_all_1999_pt_sf.rds'))){
        
           
        
        # Convert the original file into sf
        
        if(!file.exists(root_file('./1-data/4-interim/over-size-limit/prcl-all-1999-sf.rds'))){
                # Figure out which featureclass to load
                
                fp_1999 <- root_file('./1-data/3-external/manual/over-size-limit/Year2000.gdb/')
                
                # layers <- ogrListLayers(fp_1999)
                # 
                # layers[1:length(layers)]
                
                # It looks like 'parcel_area' is the spatial layer 
                
                readOGR(dsn = fp_1999,layer = 'parcel_area',verbose = FALSE,stringsAsFactors = FALSE) %>% 
                        st_as_sf() %>% 
                        st_transform(crs_proj@projargs) %>% 
                        write_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-1999-sf.rds'))
                
        }
        
        # Read in the sf object
        prcl_all_1999_sf <- read_rds(root_file('./1-data/4-interim/over-size-limit/prcl-all-1999-sf.rds'))
        
        
        # Identify the trouble polygons
        if(!file.exists(root_file('./1-data/4-interim/prcl-1999-bad-geoms.rds'))){
                prcl_all_1999_sf %>% 
                        as('Spatial') %>% 
                        clgeo_CollectionReport() %>% 
                        clgeo_SuspiciousFeatures() %>% 
                        write_rds(root_file('./1-data/4-interim/prcl-1999-bad-geoms.rds'))
        }
        
        bad_geom_indices <- read_rds(root_file('./1-data/4-interim/prcl-1999-bad-geoms.rds'))
                
        
        # Remove the trouble polygons,
        # tranform the geometries from polygon to point with `st_centroid()`,
        # and reproject
        prcl_all_1999_pt_sf <- 
                prcl_all_1999_sf[1:nrow(prcl_all_1999_sf) %!in% bad_geom_indices,] %>% 
                st_centroid() %>% 
                st_transform(crs_proj@projargs)
        
        # Save the sf object
        prcl_all_1999_pt_sf %>% write_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_1999_pt_sf.rds'))
}

prcl_all_1999_pt_sf <- read_rds(root_file('1-data/4-interim/over-size-limit/prcl_all_1999_pt_sf.rds'))

prcl_all_1999_pt_sf %>% miscgis::print_paged_sf()

```
