==> 1-data/1-notebooks/03-demo-change-inc.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r demo-chng-inc-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(scales)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(gridExtra)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(htmlTable)
library(tidyverse)
library(forcats)
library(crosstalk)
library(plotly)
library(DT)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```

```{r demo-chng-inc-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
new_spectral <- brewer.pal(11, 'Spectral')[c(3:5,8:10)]
```

```{r demo-chng-inc-objs}

# Load the map-making function
if(!exists('make_acs_map')){make_acs_map <- read_rds(root_file('./1-data/4-interim/make-acs-map-fun.rds'))}

# Load the crosstalk function
if(!exists('make_acs_crosstalk')){make_acs_crosstalk <- read_rds(root_file('./1-data/4-interim/make-acs-crosstalk-fun.rds'))}

# Load the DT function
if(!exists('make_acs_dt')){make_acs_dt <- read_rds(root_file('./1-data/4-interim/make-acs-dt-fun.rds'))}

# Load the two acs indicator objecta that are prepared for the above function
if(!exists('map_acs_data_sf')){map_acs_data_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-inds-map-sf.rds'))}

if(!exists('map_acs_lgl_data_sf')){map_acs_lgl_data_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-lgl-inds-map-sf.rds'))}

```

### Income {-}
 
This map and graph show the **change in the share of moderate/high income households** between the two observation periods for each census tract in Seattle CCD. 

The observation periods used in this method are the following five-year spans: 2005-2009 and 2011-2015.

<br>


```{r demo-chng-inc-seaccd-map}

make_acs_crosstalk(long_sf = map_acs_data_sf,
             .geog = 'tract',
             .category = 'DEMO',
             .topic = 'B19001',
             .pal = new_spectral,              
             .pal_type = 'quantile',
             .legend_title = "CHANGE IN<br>% MOD/HIGH INCOME<hr>2005-2009<br>TO<br>2011-2015",
             .height_lflt = 600,
             .height_pltly_upper = 200,
             .height_pltly_lower = 400,
             .width_pltly = 400,
             .plotly_title = "CHANGE IN % MOD/HIGH INCOME",
             zoom = 9)
```

##### COO Baselines: Seattle Subdivision vs King County {-}
```{r demo-chng-inc-comm-map, fig.height=3, fig.width=2,  fig.show='hold', dpi=150}

# SEACCD
make_acs_map(long_sf = map_acs_lgl_data_sf,
             .comms_only = TRUE,
             .baseline = 'SEACCD',
             .geog = 'tract',
             .category = 'DEMO',
             .topic = 'B19001',
             .pal = new_spectral,
             .pal_type = 'factor',
             .legend_title = "Tract<br>exceeds<br>Baseline<br>(SEACCD)",
             .css_style = paste0("float: left;","margin:3px;","text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF")) 

# KC
make_acs_map(long_sf = map_acs_lgl_data_sf,
             .comms_only = TRUE,
             .baseline = 'KC',
             .geog = 'tract',
             .category = 'DEMO',
             .topic = 'B19001',
             .pal = new_spectral,
             .pal_type = 'factor',
             .legend_title = "Tract<br>exceeds<br>Baseline<br>(KC)",
             .css_style = paste0("float: none;","margin:3px;","text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF")) 


```


##### Data: All Geographies {-}
```{r demo-chng-inc-dt}

make_acs_dt(long_sf = map_acs_data_sf,
            .category = 'DEMO',
            .topic = 'B19001',
            .col_title = "Change in Mod./High Inc. HH's")


```
