==> 1-data/1-notebooks/03-make-acs-maps.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r make-acs-maps-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(scales)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(gridExtra)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(htmlTable)
library(plotly)
library(crosstalk)
library(forcats)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```

```{r make-acs-maps-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
new_spectral <- brewer.pal(11, 'Spectral')[c(3:5,8:10)]
```

```{r make-acs-maps-data}

# Load the complete acs indicator data object
if(!exists('inds_sf')){
        inds_sf <- read_rds(root_file('./1-data/5-tidy/coo-acs-inds-not-acs-sf.rds'))
        }

# Load the complete acs popup table, 
# created in `03-make-acs-popups.Rmd`
if(!exists('tbl_popup')){
        tbl_popup <- 
                read_rds(root_file('./1-data/4-interim/acs-htmltable-popups.rds')) %>% 
                gather(TBL,POPUP,-NAME) %>% 
                mutate(CAT_TOP = str_replace_all(TBL,'_POPUP','')) %>% 
                select(-TBL)
}

# Gather the data object, join the popups, save
if(!file.exists(root_file('./1-data/4-interim/coo-acs-inds-map-sf.rds'))){
        
        joined_sf <- 
                inds_sf %>%  
                select(NAME,NAME_FULL,GEOGRAPHY,GEOG_FCT,COMMUNITY,matches('EST')) %>% 
                select(-matches('VULN_CNT_2009|VULN_PCT_2009|SEACCD|KC')) %>% 
                mutate(row = row_number()) %>% 
                gather(TBL,VAL,-NAME,-NAME_FULL,-GEOGRAPHY,-GEOG_FCT,-COMMUNITY,-geometry,-row) %>% 
                mutate(CAT = str_sub(TBL,1,4) %>% str_replace('_',''),
                       TOPIC = str_extract(TBL,'B[[:digit:]]{5}'),
                       CAT_TOP = str_c(CAT,TOPIC,sep = '_')) %>% 
                filter(str_detect(TBL, 'DEMO_CHNG_PCT|VULN_PCT')) %>% 
                st_cast('MULTIPOLYGON') %>% 
                left_join(tbl_popup, by = c('NAME','CAT_TOP')) 
        
        write_rds(joined_sf,root_file('./1-data/4-interim/coo-acs-inds-map-sf.rds'))
}

if(!file.exists(root_file('./1-data/4-interim/coo-acs-lgl-inds-map-sf.rds'))){
        
        joined_lgl_sf <- 
                inds_sf %>%  
                select(NAME,NAME_FULL,GEOGRAPHY,GEOG_FCT,COMMUNITY,matches('LGL_')) %>% 
                mutate(row = row_number()) %>% 
                gather(TBL,VAL,-NAME,-NAME_FULL,-GEOGRAPHY,-GEOG_FCT,-COMMUNITY,-geometry,-row) %>% 
                mutate(CAT = str_sub(TBL,6,10) %>% str_replace('_',''),
                       TOPIC = str_extract(TBL,'B[[:digit:]]{5}$'),
                       CAT_TOP = str_c(CAT,TOPIC,sep = '_'),
                       BASELINE = str_extract(TBL,'KC|SEACCD')) %>% 
                st_cast('MULTIPOLYGON') %>% 
                left_join(tbl_popup, by = c('NAME','CAT_TOP')) 
        
        write_rds(joined_lgl_sf,root_file('./1-data/4-interim/coo-acs-lgl-inds-map-sf.rds'))
}

if(!file.exists(root_file('./1-data/4-interim/coo-acs-lgl-typo-inds-map-sf.rds'))){
        
        joined_lgl_typo_sf <- 
                inds_sf %>%  
                select(NAME,NAME_FULL,GEOGRAPHY,GEOG_FCT,COMMUNITY,matches('TYPO.*LGL$')) %>% 
                mutate(row = row_number()) %>% 
                gather(TBL,VAL,-NAME,-NAME_FULL,-GEOGRAPHY,-GEOG_FCT,-COMMUNITY,-geometry,-row) %>% 
                mutate(CAT = str_sub(TBL,6,10) %>% str_replace('_',''),
                       TOPIC = str_extract(TBL,'B[[:digit:]]{5}$'),
                       CAT_TOP = str_c(CAT,TOPIC,sep = '_'),
                       BASELINE = str_extract(TBL,'KC|SEACCD')) %>% 
                st_cast('MULTIPOLYGON') %>% 
                left_join(tbl_popup, by = c('NAME','CAT_TOP')) 
        
        write_rds(joined_lgl_typo_sf,root_file('./1-data/4-interim/coo-acs-lgl-typo-inds-map-sf.rds'))
}

```

```{r make-acs-maps-fun}

# Make the function that will be used to build the leaflet maps for the acs indicators

if(!file.exists(root_file('./1-data/4-interim/make-acs-map-fun.rds'))){
        make_acs_map <- function(long_sf,
                         .geog,
                         .comms_only = FALSE,
                         .baseline = NULL,
                         .category,
                         .topic,
                         .pal,
                         .pal_type,
                         .legend_title,
                         zoom = 10,
                         .width = NULL,
                         .height = NULL,
                         .css_style = "text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF"){
        
        data <- 
                long_sf %>% 
                filter(GEOGRAPHY %in% .geog) %>% 
                filter(CAT %in% .category) %>% 
                filter(TOPIC %in% .topic) 
        
        if(!is.null(.baseline)){data %<>% filter(BASELINE %in% .baseline)}
        
        
        data_all_geogs <- long_sf %>% filter(CAT %in% .category & TOPIC %in% .topic)
        
        # Create outlines and labels for the COO communities
        
        comm_cnt_sf <- 
                data_all_geogs %>% 
                filter(GEOGRAPHY %in% 'community') %>% 
                mutate(geometry = st_centroid(geometry)) 
        
        comm_bounds_sf <- 
                data_all_geogs %>% 
                filter(GEOGRAPHY %in% 'community') %>%
                st_cast('LINESTRING')
        
        # Create center point for the map
        # Exclude non-COO tracts when .comms_only == TRUE
        if(.comms_only){
                data %<>% filter(!is.na(COMMUNITY))
                data_cnt <- data %>% st_union %>% st_centroid() %>% st_bbox()
        }else{
                data_cnt <- 
                        data_all_geogs %>% 
                        filter(NAME %in% 'SEACCD') %>% 
                        st_union %>% 
                        st_centroid() %>% 
                        st_bbox()
        }
        

        # Color palette
        
        if(.pal_type %in% 'quantile' & .category %in% 'DEMO'){
                
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        map_dbl(abs) %>% 
                        max(na.rm = TRUE) %>% 
                        c(-1*.,.)
                
                # create the discrete color scale
                pal <- colorBin(palette = .pal, 
                                domain = domain, 
                                bins = 6,
                                pretty = TRUE)
                lab_format <- labelFormat(suffix = '%',transform = function(x) 100 * x)
                
        }else if(.pal_type %in% 'quantile' & .category %in% 'VULN'){
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        range(na.rm = TRUE)
                
                # create the discrete color scale
                pal <- colorBin(palette = .pal, 
                                domain = domain, 
                                bins = 6,
                                pretty = TRUE)
                
                lab_format <- labelFormat(suffix = '%',transform = function(x) 100 * x)
        }else if(.pal_type %in% 'factor'){
                # Set palette domain
                domain <- data_all_geogs %>% extract2('VAL') %>% unique()
                
                # create the discrete color scale
                pal <- colorFactor(palette = .pal, 
                                domain = domain)
                lab_format <- labelFormat()
        }else {
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        range(na.rm = TRUE)
                # create the continuous color scale
                pal <- colorNumeric(.pal, domain)
                
                lab_format <- labelFormat(suffix = '%',transform = function(x) 100 * x)
        }
        
        
        # Make the leaflet map
        myLfltGrey(data,
                   width = .width,
                   height = .height) %>%
                addPolygons(fillOpacity = .75,
                            fillColor = ~pal(VAL),
                            color = col2hex('white'),
                            opacity = 1,
                            weight = .5,
                            smoothFactor = 0,
                            popup = ~POPUP,
                            popupOptions = popupOptions(maxWidth = 1000)) %>%
                addPolylines(data = comm_bounds_sf,
                             fillOpacity = 0,
                             color = proj_dark_grey,
                             opacity = 1,
                             weight = 2,
                             smoothFactor = 0) %>%
                addAwesomeMarkers(data = comm_cnt_sf,
                                  options = markerOptions(opacity = 0),
                                  label = ~ NAME_FULL, 
                                  labelOptions = labelOptions(noHide = TRUE,
                                                              offset = c(0,0),
                                                              direction = 'auto', 
                                                              textOnly = TRUE)) %>% 
                addLegend(title = .legend_title,
                          position = 'topright',
                          pal = pal,
                          values = domain,
                          opacity = .75,
                          labFormat = lab_format) %>%
                setView(lng = data_cnt[['xmin']],
                        lat = data_cnt[['ymin']],
                        zoom = zoom) %>% 
                styleWidget(style = .css_style) %>% 
                myLfltOpts(hideControls = FALSE) %>% 
                addFullscreenControl(pseudoFullscreen = TRUE)
        
        
        
}
        
        make_acs_map %>% write_rds(root_file('./1-data/4-interim/make-acs-map-fun.rds'))
        
}

```

```{r make-acs-crosstalk-fun}
if(!file.exists(root_file('./1-data/4-interim/make-acs-crosstalk-fun.rds'))){
        make_acs_crosstalk <- function(long_sf,
                                 .geog,
                                 .comms_only = FALSE,
                                 .category,
                                 .topic,
                                 .pal,
                                 .pal_type,
                                 .legend_title,
                                 zoom = 10,
                                 .height_lflt = 400,
                                 .width_lflt = '100%',
                                 .css_style = "text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF",
                                 .height_pltly_upper = 200,
                                 .height_pltly_lower = 200,
                                 .width_pltly = '100%',
                                 .plotly_title){
        
        # browser()
        
        data <- 
                long_sf %>% 
                filter(GEOGRAPHY %in% .geog) %>% 
                filter(CAT %in% .category) %>% 
                filter(TOPIC %in% .topic) %>% 
                mutate(NAME = fct_reorder(NAME,VAL,.desc = TRUE),
                        COLOR = case_when(.$COMMUNITY %in% 'RV' ~ "#1F77B4",
                                         .$COMMUNITY %in% 'WC' ~ "#2CA02C",
                                         .$COMMUNITY %in% 'STKW' ~ "#D62728",
                                         TRUE ~ 'grey50')) %>% 
                arrange(NAME)
        
        data_pt_sf <- data %>% st_centroid()
        
        data_poly_sf <- data 
        
        data_all_geogs <- long_sf %>% filter(CAT %in% .category & TOPIC %in% .topic)
        
        # Create outlines and labels for the COO communities

        comm_cnt_sf <- 
                data_all_geogs %>% 
                filter(GEOGRAPHY %in% 'community') %>% 
                mutate(geometry = st_centroid(geometry)) 
        
        comm_bounds_sf <- 
                data_all_geogs %>% 
                filter(GEOGRAPHY %in% 'community') %>%
                st_cast('LINESTRING')
        
        # Create center point for the map
        # Exclude non-COO tracts when .comms_only == TRUE
        if(.comms_only){
                data %<>% filter(!is.na(COMMUNITY))
                data_cnt <- data %>% st_union %>% st_centroid() %>% st_bbox()
        }else{
                data_cnt <- 
                        data_all_geogs %>% 
                        filter(NAME %in% 'SEACCD') %>% 
                        st_union %>% 
                        st_centroid() %>% 
                        st_bbox()
        }
 
        # Color palette
        
        if(.pal_type %in% 'quantile' & .category %in% 'DEMO'){
                
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        map_dbl(abs) %>% 
                        max(na.rm = TRUE) %>% 
                        c(-1*.,.)
                
                # create the discrete color scale
                pal <- colorBin(palette = .pal, 
                                domain = domain, 
                                bins = 6,
                                pretty = TRUE)
                
        }else if(.pal_type %in% 'quantile' & .category %in% 'VULN'){
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        range(na.rm = TRUE)
                
                # create the discrete color scale
                pal <- colorBin(palette = .pal, 
                                domain = domain, 
                                bins = 6,
                                pretty = TRUE)
        }else{
                # Set palette domain
                domain <- 
                        data_all_geogs %>% 
                        extract2('VAL') %>% 
                        range(na.rm = TRUE)
                # create the continuous color scale
                pal <- colorNumeric(.pal, domain)
        }
        

        
        # Shared data
                     
        shared_data <- SharedData$new(data_pt_sf)
        
        
        lflt_map <-            
                myLfltGrey(shared_data,
                           width = .width_lflt,
                           height = .height_lflt) %>%
                addLabelOnlyMarkers(label = '') %>% 
                addPolygons(data = data_poly_sf,
                            fillOpacity = .75,
                            fillColor = ~pal(VAL),
                            color = col2hex('white'),
                            opacity = 1,
                            weight = .5,
                            smoothFactor = 0,
                            popup = ~POPUP,
                            popupOptions = popupOptions(maxWidth = 1000)) %>%
                addPolylines(data = comm_bounds_sf,
                             fillOpacity = 0,
                             color = proj_dark_grey,
                             opacity = 1,
                             weight = 2,
                             smoothFactor = 0) %>%
                addAwesomeMarkers(data = comm_cnt_sf,
                                  options = markerOptions(opacity = 0),
                                  label = ~ NAME_FULL, 
                                  labelOptions = labelOptions(noHide = TRUE,
                                                              offset = c(0,0),
                                                              direction = 'auto', 
                                                              textOnly = TRUE)) %>% 
                addLegend(title = .legend_title,
                          position = 'topright',
                          pal = pal,
                          values = domain,
                          opacity = .75,
                          labFormat = labelFormat(suffix = '%',transform = function(x) 100 * x)) %>%
                setView(lng = data_cnt[['xmin']],
                        lat = data_cnt[['ymin']],
                        zoom = zoom) %>% 
                styleWidget(style = .css_style) %>% 
                myLfltOpts(hideControls = FALSE) %>% 
                addFullscreenControl(pseudoFullscreen = TRUE)
                
        # Aggregate geoms
        data_agg_geogs <- 
                long_sf %>% 
                filter(GEOGRAPHY %!in% 'tract') %>% 
                filter(CAT %in% .category) %>% 
                filter(TOPIC %in% .topic) %>% 
                mutate(NAME = fct_reorder(NAME,as.numeric(GEOG_FCT),.desc = FALSE),
                       COLOR = case_when(.$COMMUNITY %in% 'RV' ~ "#1F77B4",
                                         .$COMMUNITY %in% 'WC' ~ "#2CA02C",
                                         .$COMMUNITY %in% 'STKW' ~ "#D62728",
                                         TRUE ~ 'grey50')) %>% 
                arrange(NAME)
        
        
        # Plots
        
        if(.category %in% 'VULN'){
                plotly_range <- list(0,1)
        }else{
                plotly_range <- 
                        long_sf %>% 
                        filter(CAT %in% 'DEMO') %>% 
                        filter(TOPIC %in% 'B03002') %>% 
                        extract2('VAL') %>% 
                        abs %>% 
                        max %>% 
                        plyr::round_any(.1) %>% 
                        list(multiply_by(.,-1),.)
                        }
        
        upper <- 
                plot_ly(data_agg_geogs,
                                x = ~VAL, 
                                y=  ~NAME, 
                                color = ~NAME, 
                                colors = ~COLOR, 
                                type = 'scatter', 
                                mode = 'markers',
                                height = .height_pltly_upper, 
                                width = .width_pltly) %>% 
                                layout(title = .plotly_title,
                                       xaxis = list(range = plotly_range, tickformat = '%', title = ''),
                                       yaxis = list(title = "")) %>% 
                hide_legend()

        lower <-
                plot_ly(shared_data, 
                                x = ~VAL, 
                                y=  ~NAME, 
                                color = ~NAME, 
                                colors = ~COLOR, 
                                type = 'scatter', 
                                mode = 'markers', 
                                height = .height_pltly_lower, 
                                width = .width_pltly) %>% 
                                layout(
                                        xaxis = list(range = plotly_range, tickformat = '%', title = ''),
                                        yaxis = list(title = "TRACTS")
                                ) %>% 
                hide_legend()
        
        # Crosstalk
        
        bscols(
                widths = c(5, 7),
                lflt_map,
                list(upper,lower)
                
                )
        
        
        }
        
        make_acs_crosstalk %>% write_rds(root_file('./1-data/4-interim/make-acs-crosstalk-fun.rds'))
        
}
```

```{r make-acs-dt-fun}

if(!file.exists(root_file('./1-data/4-interim/make-acs-dt-fun.rds'))){
        make_acs_dt <- function(long_sf,
                                 .category,
                                 .topic,
                                .col_title = ''){
        
        # browser()
        
        data <- 
                long_sf %>% 
                filter(CAT %in% .category) %>% 
                filter(TOPIC %in% .topic) %>% 
                mutate(NAME_FULL = str_replace(NAME_FULL,', King County, Washington',''),
                       VAL = scales::percent(VAL)) %>% 
                unclass %>% 
                as_tibble %>% 
                select(NAME_FULL,COMMUNITY,GEOGRAPHY, TBL, VAL) %>% 
                spread(TBL,VAL) 
                
        
        data %>% datatable(rownames = FALSE,
                           filter = 'top', 
                           colnames = c(
                                   'Name', 
                                   'Community',
                                   'Geog. Type',
                                   .col_title
                           ))
                
        
        
        }
        
        make_acs_dt %>% write_rds(root_file('./1-data/4-interim/make-acs-dt-fun.rds'))
        
}

```

```{r make-acs-maps-examples, eval=FALSE, include = FALSE}


make_acs_map <- read_rds(root_file('./1-data/4-interim/make-acs-map-fun.rds'))

map_data_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-inds-map-sf.rds'))

make_acs_map(long_sf = map_data_sf,
             .geog = 'tract',
             .category = 'VULN',
             .topic = 'B03002',
             .pal = 'plasma',
             .pal_type = 'quantile',
             .legend_title = "SHARE OF<br>PEOPLE OF COLOR<hr>2011-2015")

make_acs_map(long_sf = map_data_sf,
             .geog = 'tract',
             .category = 'DEMO',
             .topic = 'B03002',
             .pal = new_spectral,
             .pal_type = 'quantile',
             .legend_title = "CHANGE IN<br>SHARE OF<br>PEOPLE OF COLOR<hr>2011-2015")

make_acs_map(long_sf = map_data_sf,
             .comms_only = TRUE,
             .geog = 'tract',
             .category = 'VULN',
             .topic = 'B15002',
             .pal = 'plasma',
             .legend_title = "SHARE OF<br>PEOPLE OF COLOR<hr>2011-2015")

make_acs_map(long_sf = map_data_sf,
             .geog = 'community',
             .category = 'DEMO',
             .topic = 'B03002',
             .pal = div_pal,
             .legend_title = "CHANGE IN SHARE<br>PEOPLE OF COLOR<hr>2009-2015")

```
