==> 1-data/1-notebooks/03-make-acs-sparklines.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r make-acs-sparklines-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(sparkline)
library(scales)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(gridExtra)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(htmlTable)
library(plotly)
library(crosstalk)
library(forcats)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```

```{r make-acs-sparklines-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
new_spectral <- brewer.pal(11, 'Spectral')[c(3:5,8:10)]
```

```{r make-acs-sparklines}

if(!file.exists(root_file('./1-data/4-interim/acs-sparklines.rds'))){
        
        # store values for each acs indicator in a list column
        nested_data <- 
                read_rds(root_file('./1-data/5-tidy/coo-acs-inds-not-acs-sf.rds')) %>%  
                select(NAME,NAME_FULL,GEOGRAPHY,GEOG_FCT,COMMUNITY,SEACCD_LGL,matches('EST')) %>% 
                select(-matches('VULN_CNT_2009|VULN_PCT_2009|SEACCD_B|KC')) %>% 
                mutate(row = row_number()) %>% 
                gather(TBL,VAL,-NAME,-NAME_FULL,-GEOGRAPHY,-GEOG_FCT,-COMMUNITY,-SEACCD_LGL,-geometry,-row) %>% 
                mutate(CAT = str_sub(TBL,1,4) %>% str_replace('_',''),
                       TOPIC = str_extract(TBL,'B[[:digit:]]{5}'),
                       CAT_TOP = str_c(CAT,TOPIC,sep = '_')) %>% 
                filter(str_detect(TBL, 'DEMO_CHNG_PCT|VULN_PCT')) %>%
                group_by(CAT_TOP) %>% 
                nest %>% 
                crossing('SEACCD_ONLY_LGL' = c(TRUE,FALSE))
        
        # Pass the nested data and the SEACCD_ONLY_LGL column to sparkline()
        
        my_fun <- function(data){
                browser()
        }
        
        
        mutate(nested_data, TEST = map(data, my_fun))
        filter(NAME %in% 'SEACCD') %>% extract2('VAL')
        
        sparkline_tbl <- 
                mutate(nested_data,
                       SPRK_VALS = map2(.x = data, 
                                        .y = SEACCD_ONLY_LGL,
                                        .f = ~ .x %>%  {if(.y) filter(., SEACCD_LGL) else .} %>%  extract2('VAL')),
                       BASELINE = map2(.x = data, 
                                       .y = SEACCD_ONLY_LGL,
                                       .f = ~ if_else(.y, 
                                                      filter(.x, NAME %in% 'SEACCD') %>% extract2('VAL'),
                                                      filter(.x, NAME %in% 'KC') %>% extract2('VAL')))) %>% 
                mutate(
                        SPARKLINE = map2(
                                .x = SPRK_VALS, 
                                .y = BASELINE,
                                .f = ~ sparkline( values = .x,
                                                  type = 'box', 
                                                  lineColor = proj_dark_grey, 
                                                  whiskerColor = proj_dark_grey, 
                                                  target = .y,
                                                  targetColor = 'tomato',
                                                  outlierFillColor = 'white', 
                                                  outlierLineColor = proj_dark_grey, 
                                                  boxFillColor = proj_light_grey,
                                                  boxLineColor = proj_dark_grey, 
                                                  medianColor = 'none',
                                                  tooltipFormatFieldlist = c('med', 'lq', 'uq'), 
                                                  tooltipFormatFieldlistKey =  'field'))
                )
                       
        
        # prepare a named vector to match topics with topic table id's
        acs_tbls <- 
                c('race', 'housing', 'ed', 'income') %>% 
                set_names(c('B03002','B15002', 'B19001', 'B25033'))
        
        # Create the NAME column
        sprk_tbl <- 
                sparkline_tbl %>% 
                separate(CAT_TOP,c('CAT', 'TOPIC_CODE'), sep = '_') %>% 
                mutate(TOPIC = map_chr(TOPIC_CODE, ~acs_tbls[[.x]]),
                       BASELINE_GEOG = if_else(SEACCD_ONLY_LGL, 'SEACCD','KC'),
                       NAME = str_c(TOPIC, CAT, BASELINE_GEOG, sep = '_'))
        
        # Make the named vector of sparkline html widgets
        sprk <- set_names(x = sprk_tbl$SPARKLINE, nm = sprk_tbl$NAME)
        
        # Save the object
        write_rds(sprk,
                  root_file('./1-data/4-interim/acs-sparklines.rds'))
}

```
