==> 1-data/1-notebooks/03-make-hous-val-inds.Rmd <==
---
df_print: tibble
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
editor_options:
  chunk_output_type: console
---

```{r hous-val-inds-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(gdalUtils)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(foreign)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(maptools)
library(cleangeo)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```

```{r hous-val-inds-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
```

```{r hous-val-inds-make}

if(!file.exists(root_file('./1-data/4-interim/coo-hous-val-inds.rds'))){
        
        # Load the long parcel data
        
        val_long <- read_rds(root_file('./1-data/4-interim/hous-val-long.rds'))
        
        
        # Load the COO geoms data (all inds will be joined to this object)
        
        geoms_sf <- 
                read_rds(root_file('./1-data/4-interim/acs_norml_all_geoms_sf.rds')) %>%
                select(NAME:GEOG_FCT) %>% 
                mutate(JOIN_ID = if_else(is.na(GEOID6),NAME,GEOID6)) %>% 
                select(JOIN_ID,everything(),-NAME)
        
        # Ratio indicator -------------------------------------------------------------------------------------------------
        # -----------------------------------------------------------------------------------------------------------------
        
        get_comparison <- function(df,.comp_col,.val_col){
                
                df %>% filter(NAME %in% .comp_col) %>% extract2(.val_col) %>% as.integer
        }
        
        val_ratio <-         
                val_long %>% 
                group_by(YEAR) %>% 
                nest() %>% 
                mutate(TOTAL_VAL_MEDIAN_SEACCD = map_int(.x = data, 
                                                         .f = get_comparison, 
                                                         .comp_col = 'SEACCD', 
                                                         .val_col = 'TOTAL_VAL_MEDIAN'),
                       TOTAL_VAL_MEDIAN_KC = map_int(.x = data, 
                                                     .f = get_comparison, 
                                                     .comp_col = 'KC', 
                                                     .val_col = 'TOTAL_VAL_MEDIAN')) %>% 
                unnest(data,.drop = FALSE) %>% 
                mutate(RATIO_SEACCD = case_when(
                        !.$SEACCD_LGL | is.na(SEACCD_LGL) ~ NA_real_,
                        TRUE ~ divide_by(.$TOTAL_VAL_MEDIAN,.$TOTAL_VAL_MEDIAN_SEACCD) %>% round_any(.01)
                ),
                RATIO_KC = divide_by(TOTAL_VAL_MEDIAN,TOTAL_VAL_MEDIAN_KC) %>% round_any(.01)) %>%
                group_by(YEAR) %>% 
                mutate(TYPE_SEACCD = if_else(ntile(RATIO_SEACCD,5) > 3,'HIGH','LOW/MOD'),
                       TYPE_KC = if_else(ntile(RATIO_KC,5) > 3,'HIGH','LOW/MOD')) %>% 
                ungroup %>% 
                mutate(KEY_MEDIAN = str_c('VAL_MEDIAN',YEAR,sep = '_'),
                       KEY_MEDIAN_SEACCD = str_c('VAL_MEDIAN_SEACCD',YEAR,sep = '_'),
                       KEY_MEDIAN_KC = str_c('VAL_MEDIAN_KC',YEAR,sep = '_'),
                       KEY_RATIO_SEACCD = str_c('VAL_RATIO_SEACCD',YEAR,sep = '_'),
                       KEY_RATIO_KC = str_c('VAL_RATIO_KC',YEAR,sep = '_'),
                       KEY_TYPE_SEACCD = str_c('VAL_TYPE_SEACCD',YEAR,sep = '_'),
                       KEY_TYPE_KC = str_c('VAL_TYPE_KC',YEAR,sep = '_')) %>% 
                select(matches('NAME|TOTAL|RATIO|TYPE|KEY')) %>% 
                spread(KEY_MEDIAN,TOTAL_VAL_MEDIAN) %>% 
                spread(KEY_MEDIAN_SEACCD, TOTAL_VAL_MEDIAN_SEACCD) %>% 
                spread(KEY_MEDIAN_KC, TOTAL_VAL_MEDIAN_KC) %>% 
                spread(KEY_RATIO_SEACCD,RATIO_SEACCD) %>% 
                spread(KEY_RATIO_KC,RATIO_KC) %>% 
                spread(KEY_TYPE_SEACCD,TYPE_SEACCD) %>% 
                spread(KEY_TYPE_KC,TYPE_KC) %>% 
                group_by(NAME) %>% 
                summarise_all(first_not_na) %>% 
                select(JOIN_ID = NAME,everything(),-NAME_FULL)
        
        # High value neighbors --------------------------------------------------------------------------------------------
        # -----------------------------------------------------------------------------------------------------------------
        
        all_geoms_val_ratio_2015 <- 
                geoms_sf %>% 
                data.frame() %>% 
                left_join(val_ratio, by = 'JOIN_ID') %>%  
                select(JOIN_ID,VAL_TYPE_SEACCD_2015,VAL_TYPE_KC_2015,geometry) %>% 
                st_as_sf
                
        tr_sf <- 
                geoms_sf %>% 
                data.frame() %>% 
                left_join(val_ratio, by = 'JOIN_ID') %>%  
                filter(GEOGRAPHY %in% 'tract') %>% 
                select(JOIN_ID,VAL_TYPE_SEACCD_2015,VAL_TYPE_KC_2015,geometry) %>% 
                st_as_sf
        
        
        adj_to_high <- function(nb,full_obj,type,.type_col){
                # browser()
                if(type %in% 'HIGH'){return(as.logical(NA))}
                
                full_obj[nb,] %>% extract2(.type_col) %>% map_lgl(~ .x %in% 'HIGH') %>% any
        }
        
        val_adj <- 
                tr_sf %>% 
                mutate(NB = st_intersects(geometry, geometry)) %>% 
                mutate(NB_GEOMS = map(NB, ~ st_geometry(tr_sf[.x,]))) %>% 
                mutate(GEOM_LST = map(geometry, ~ st_geometry(.x) %>% st_set_crs(st_crs(tr_sf)))) %>% 
                mutate(RELATE = map2(.x = NB_GEOMS, .y = GEOM_LST,.f = ~ st_relate(.y,.x))) %>% 
                mutate(RELATE_LGL = map(RELATE, str_detect, pattern = '^[[:alnum:]]{4}1')) %>% 
                mutate(NB_ROOK = map2(.x = NB, .y = RELATE_LGL, ~.x[.y])) %>% 
                mutate(ADJ_SEACCD_LGL = map2_lgl(.x = NB_ROOK,
                                  .y = VAL_TYPE_SEACCD_2015,
                                  .f = adj_to_high,
                                  full_obj = tr_sf,
                                  .type_col = 'VAL_TYPE_SEACCD_2015'),
                       ADJ_KC_LGL = map2_lgl(.x = NB_ROOK,
                                  .y = VAL_TYPE_KC_2015,
                                  .f = adj_to_high,
                                  full_obj = tr_sf,
                                  .type_col = 'VAL_TYPE_KC_2015')) %>% 
                data.frame() %>% 
                select(JOIN_ID,ADJ_SEACCD_LGL,ADJ_KC_LGL, NEIGHBORS = NB_ROOK) %>% 
                left_join(data.frame(geoms_sf),., by = 'JOIN_ID') %>% 
                data.frame() %>% 
                select(JOIN_ID, matches('ADJ'),NEIGHBORS)
                
                

        # Rate indicator --------------------------------------------------------------------------------------------------
        # -----------------------------------------------------------------------------------------------------------------
        find_appr_rate <- function(df, .start_year, .end_year, .seaccd_only = FALSE){
                
                if(df %>% extract2('TOTAL_VAL_MEDIAN') %>% map_lgl(is.na) %>% all){
                        rate <- NA_real_
                }else{
                        start <- df %>% filter(YEAR %in% .start_year) %>% extract2('TOTAL_VAL_MEDIAN')
                        
                        end <- df %>% filter(YEAR %in% .end_year) %>% extract2('TOTAL_VAL_MEDIAN')
                        
                        rate <- divide_by(end,start) %>% round_any(.01)
                        
                        if(.seaccd_only & !isTRUE(unique(extract2(df,'SEACCD_LGL')))) {rate <- NA_real_}
                }
                return(rate)
        }
        
        val_appr_rate <- 
                val_long %>% 
                group_by(NAME) %>%
                nest() %>% 
                mutate(APPR_RATE_99_04_SEACCD = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,1999,2004, .seaccd_only = TRUE)),
                       APPR_RATE_04_15_SEACCD = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,2004,2015, .seaccd_only = TRUE)),
                       APPR_RATE_99_15_SEACCD = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,1999,2015, .seaccd_only = TRUE)),
                       APPR_RATE_99_04_KC = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,1999,2004,.seaccd_only = FALSE)),
                       APPR_RATE_04_15_KC = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,2004,2015,.seaccd_only = FALSE)),
                       APPR_RATE_99_15_KC = map_dbl(.x = data,.f =  ~ find_appr_rate(.x,1999,2015,.seaccd_only = FALSE))) %>% 
                mutate(APPR_TYPE_99_04_SEACCD = if_else(ntile(APPR_RATE_99_04_SEACCD,5) > 3,'HIGH','LOW/MOD'),
                       APPR_TYPE_04_15_SEACCD = if_else(ntile(APPR_RATE_04_15_SEACCD,5) > 3,'HIGH','LOW/MOD'),
                       APPR_TYPE_99_15_SEACCD = if_else(ntile(APPR_RATE_99_15_SEACCD,5) > 3,'HIGH','LOW/MOD'),
                       APPR_TYPE_99_04_KC = if_else(ntile(APPR_RATE_99_04_KC,5) > 3,'HIGH','LOW/MOD'),
                       APPR_TYPE_04_15_KC = if_else(ntile(APPR_RATE_04_15_KC,5) > 3,'HIGH','LOW/MOD'),
                       APPR_TYPE_99_15_KC = if_else(ntile(APPR_RATE_99_15_KC,5) > 3,'HIGH','LOW/MOD')) %>% 
                unnest(data,.drop = FALSE) %>% 
                group_by(NAME) %>% 
                summarise_all(first_not_na) %>% 
                mutate(JOIN_ID = NAME) %>% 
                select(JOIN_ID,matches('APPR'))
        
        
        # Join, create final inds, and save  --------------------------------------------------------------------------------------------------
        # -----------------------------------------------------------------------------------------------------------------
        
        hous_val_inds_sf <- 
                list(val_ratio, val_adj, val_appr_rate) %>% 
                map(~ .x %>% data.frame() %>% select( -matches('geom'))) %>% 
                prepend(list(data.frame(geoms_sf))) %>% 
                reduce(left_join, by = 'JOIN_ID') %>% 
                mutate(TYPO_HOUS_ADJ_SEACCD_LGL = if_else(VAL_TYPE_SEACCD_2015 %in% 'LOW/MOD' &
                                                             APPR_TYPE_04_15_SEACCD %in% 'LOW/MOD' &
                                                             ADJ_SEACCD_LGL,
                                                     TRUE,
                                                     FALSE),
                       TYPO_HOUS_ADJ_KC_LGL = if_else(VAL_TYPE_KC_2015 %in% 'LOW/MOD' &
                                                             APPR_TYPE_04_15_KC %in% 'LOW/MOD' &
                                                             ADJ_KC_LGL,
                                                     TRUE,
                                                     FALSE),
                       TYPO_HOUS_ACCL_SEACCD_LGL = if_else(VAL_TYPE_SEACCD_2015 %in% 'LOW/MOD' &
                                                              APPR_TYPE_04_15_SEACCD %in% 'HIGH',
                                                      TRUE,
                                                      FALSE),
                       TYPO_HOUS_ACCL_KC_LGL = if_else(VAL_TYPE_KC_2015 %in% 'LOW/MOD' &
                                                              APPR_TYPE_04_15_KC %in% 'HIGH',
                                                      TRUE,
                                                      FALSE),
                       TYPO_HOUS_APPR_SEACCD_LGL = if_else(VAL_TYPE_SEACCD_1999 %in% 'LOW/MOD' & 
                                                              VAL_TYPE_SEACCD_2015 %in% 'HIGH' & 
                                                              APPR_TYPE_99_15_SEACCD %in% 'HIGH',
                                                      TRUE,
                                                      FALSE),
                       TYPO_HOUS_APPR_KC_LGL = if_else(VAL_TYPE_KC_1999 %in% 'LOW/MOD' & 
                                                              VAL_TYPE_KC_2015 %in% 'HIGH' & 
                                                              APPR_TYPE_99_15_KC %in% 'HIGH',
                                                      TRUE,
                                                      FALSE)) %>% 
                st_as_sf 
        
        
        write_rds(hous_val_inds_sf, root_file('./1-data/4-interim/coo-hous-val-inds.rds'))
        

}


```
