==> 1-data/1-notebooks/03-vuln-intro.Rmd <==
---
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

\cleardoublepage

```{r vuln-intro-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(sparkline)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(leaflet.extras)
library(viridisLite)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
```

```{r vuln-intro-colors}
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_4 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
pal_rgb_6 <- miscgis::miscgis_pals$tableau_cat[c("red","gold","green","blue","orange","purple")] %>% unlist %>% palette()
proj_light_grey <- col2hex("grey75")
proj_grey <- col2hex("grey50")
proj_dark_grey <- col2hex("grey25")
plasma_bivariate <- viridisLite::plasma(n = 9)[c(3,7)]
```


```{r vuln-intro-load-objs}

# Load census indicators as an sf object
if(!exists('census_inds_sf')){census_inds_sf <- read_rds(root_file('./1-data/5-tidy/coo-acs-inds-not-acs-sf.rds'))}

inds_seaccd <- 
        crossing(c('VULN_PCT_2015_SEACCD','DEMO_CHNG_PCT_SEACCD'),
                c('B03002_EST','B15002_EST', 'B19001_EST', 'B25033_EST')) %>% 
        transpose() %>% 
        map_chr(str_c, collapse = '_') %>% 
        tibble('IND' = .,
               'TOPIC' = c('race', 'housing', 'ed', 'income') %>% rep(2),
               'CATEGORY' = c(rep('DEMO_CHNG',4),rep('VULN',4))) %>% 
        mutate(VALUE = map_dbl(IND, ~ census_inds_sf %>% unclass %>% as_tibble %>% select(matches(.x)) %>% slice(1) %>% extract2(1)),
               VALUE_CHR = VALUE %>% multiply_by(100) %>% round_any(1) %>% str_c('%'),
               NAME = str_c(TOPIC,CATEGORY,sep = '_')) %>% 
               {set_names(x = .$VALUE_CHR, nm = .$NAME)}

inds_kc <- 
        crossing(c('VULN_PCT_2015_KC','DEMO_CHNG_PCT_KC'),
                c('B03002_EST','B15002_EST', 'B19001_EST', 'B25033_EST')) %>% 
        transpose() %>% 
        map_chr(str_c, collapse = '_') %>% 
        tibble('IND' = .,
               'TOPIC' = c('race', 'housing', 'ed', 'income') %>% rep(2),
               'CATEGORY' = c(rep('DEMO_CHNG',4),rep('VULN',4))) %>% 
        mutate(VALUE = map_dbl(IND, ~ census_inds_sf %>% unclass %>% as_tibble %>% select(matches(.x)) %>% slice(1) %>% extract2(1)),
               VALUE_CHR = VALUE %>% multiply_by(100) %>% round_any(1) %>% str_c('%'),
               NAME = str_c(TOPIC,CATEGORY,sep = '_')) %>% 
               {set_names(x = .$VALUE_CHR, nm = .$NAME)}

# Load the sparklines

if(!exists('sprk')){ sprk <- read_rds(root_file('./1-data/4-interim/acs-sparklines.rds'))

        }


# Load the map-making function
if(!exists('make_acs_map')){ make_acs_map <- read_rds(root_file('./1-data/4-interim/make-acs-map-fun.rds')) }

# Load the acs indicator object that is prepared for the above function 

if(!exists('map_acs_lgl_typo_data_sf')){map_acs_lgl_typo_data_sf <- read_rds(root_file('./1-data/4-interim/coo-acs-lgl-typo-inds-map-sf.rds'))}


```
 
The concept of "vulnerability" has received considerable attention by social science researchers, policy experts, and the media; it has also been subject to some criticism [@cardona_need_2004].

A tract is considered "vulnerable" if it exceeds the comparison population for any **three of the four** indicators:

##### SEACCD {-}
  * Greater than **`r inds_seaccd[['race_VULN']]`** of the population belong to communities of color `r sprk[['race_VULN_SEACCD']]`
  * Greater than **`r inds_seaccd[['housing_VULN']]`** of households are renters `r sprk[['housing_VULN_SEACCD']]`
  * Greater than **`r inds_seaccd[['ed_VULN']]`** of the population 25 years and older do not have a bachelor’s degree `r sprk[['ed_VULN_SEACCD']]`
  * Greater than **`r inds_seaccd[['income_VULN']]`** of households have incomes at or below 80% of the HUD-adjusted median family income (MFI)^[HUD's FY2015 MFI for the Seattle-Bellevue, WA HUD Metro FMR Area was $89,500; this does *not* account for household size.] `r sprk[['income_VULN_SEACCD']]`

##### King County {-}
  * Greater than **`r inds_kc[['race_VULN']]`** of the population belong to communities of color `r sprk[['race_VULN_KC']]`
  * Greater than **`r inds_kc[['housing_VULN']]`** of households are renters `r sprk[['housing_VULN_KC']]`
  * Greater than **`r inds_kc[['ed_VULN']]`** of the population 25 years and older do not have a bachelor’s degree `r sprk[['ed_VULN_KC']]`
  * Greater than **`r inds_kc[['income_VULN']]`** of households have incomes at or below 80% of the HUD-adjusted median family income (MFI) `r sprk[['income_VULN_KC']]`

  
##### COO Baselines: Seattle Subdivision vs King County {-}
```{r vuln-lgl-map, fig.height=3, fig.width=2,  fig.show='hold', dpi=150}

# SEACCD
make_acs_map(long_sf = map_acs_lgl_typo_data_sf,
             .comms_only = TRUE,
             .baseline = 'SEACCD',
             .geog = 'tract',
             .category = 'VULN',
             .topic = NA,
             .pal = plasma_bivariate,
             .pal_type = 'factor',
             .legend_title = "Vulnerable<br>Tract<br>(SEACCD)",
             .css_style = paste0("float: left;","margin:3px;","text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF")) 

# KC
make_acs_map(long_sf = map_acs_lgl_typo_data_sf,
             .comms_only = TRUE,
             .baseline = 'KC',
             .geog = 'tract',
             .category = 'VULN',
             .topic = NA,
             .pal = plasma_bivariate,
             .pal_type = 'factor',
             .legend_title = "Vulnerable<br>Tract<br>(KC)",
             .css_style = paste0("float: none;","margin:3px;","text-transform:uppercase;text-shadow:-1px 0 #FFFFFF,0 1px #FFFFFF,1px 0 #FFFFFF,0 -1px #FFFFFF")) 


```
